+++ CHANGELOG.md
@@ -9,6 +9,10 @@ nav_order: 1
 <!-- Always keep the following header in place: -->
 <!-- [MAJOR.MINOR.PATCH] - YYYY-MM-DD -->
 
+## [MAJOR.MINOR.PATCH] - YYYY-MM-DD
+
+- Fix `ip_filter` conversion issue
+
 ## [4.18.0] - 2024-05-22
 
 - Fix `ip_filter`, `ip_filter_string`, and `ip_filter_object` crash when receive an unexpected type